Added validation to dtos and controllers

See merge request Jelena/task9_university!12

+++ src/main/java/com/foxminded/controllers/CourseController.java
@@ -9,6 +9,7 @@ import com.foxminded.services.exceptions.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,6 +17,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
+import javax.validation.Valid;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Pattern;
 
 @Controller
 @RequestMapping("/courses")
@@ -24,6 +28,7 @@ public class CourseController {
     private final CourseService service;
     private final UserCourseService userCourseService;
     private static final String REDIRECT_COURSES = "redirect:/courses";
+    private static final String REDIRECT_NEW = "redirect:/courses/new";
     private static final String COURSE = "course";
     private static final String USER = "user";
     private static final String MESSAGE = "message";
@@ -67,26 +72,32 @@ public class CourseController {
     }
 
     @PostMapping("/new")
-    public String createCourse(@ModelAttribute(COURSE) CourseDto course,
-                               HttpSession session,
+    public String createCourse(@Valid @ModelAttribute(COURSE) CourseDto course,
+                               BindingResult bindingResult,
                                RedirectAttributes redirectAttributes) {
+        if (bindingResult.hasErrors()) {
+            redirectAttributes.addFlashAttribute(MESSAGE,
+                    bindingResult.getAllErrors().get(0).getDefaultMessage());
+            return REDIRECT_NEW;
+        }
         try {
             service.add(course);
         } catch (EntityAlreadyExistsException e) {
             redirectAttributes.addFlashAttribute(MESSAGE, "Course already exists.");
         }
-        try {
-            userCourseService.saveUserForCourse((UserDto)session.getAttribute(USER), course);
-        } catch (EntityAlreadyExistsException e) {
-            redirectAttributes.addFlashAttribute(MESSAGE, "User already on course.");
-        }
         return REDIRECT_COURSES;
     }
 
     @PostMapping("/add")
-    public String add(@ModelAttribute(COURSE) CourseDto course,
+    public String add(@Valid @ModelAttribute(COURSE) CourseDto course,
+                      BindingResult bindingResult,
                       HttpSession session,
                       RedirectAttributes redirectAttributes) {
+        if (bindingResult.hasErrors()) {
+            redirectAttributes.addFlashAttribute(MESSAGE,
+                    bindingResult.getAllErrors().get(0).getDefaultMessage());
+            return REDIRECT_NEW;
+        }
         try {
             userCourseService.saveUserForCourse((UserDto)session.getAttribute(USER), course);
         } catch (EntityAlreadyExistsException e) {
@@ -96,24 +107,39 @@ public class CourseController {
     }
 
     @PostMapping("/remove")
-    public String remove(@ModelAttribute("course_remove") String course,
+    public String remove(@ModelAttribute("course_remove")
+                             @NotBlank(message = "Course name can't be blank.")
+                             @Pattern(regexp="^[A-Za-z ]*$", message = "Course name should be in latin alphabet.")
+                                     String course,
+                         BindingResult bindingResult,
                          HttpSession session,
                          RedirectAttributes redirectAttributes) {
-        try {
-            userCourseService.deleteUserForCourse((UserDto)session.getAttribute(USER), service.findByName(course));
-        } catch (EntityNotFoundException e) {
-            redirectAttributes.addFlashAttribute(MESSAGE, "User not on the course.");
+        if (bindingResult.hasErrors()) {
+            redirectAttributes.addFlashAttribute(MESSAGE,
+                    bindingResult.getAllErrors().get(0).getDefaultMessage());
+        } else {
+            try {
+                userCourseService.deleteUserForCourse((UserDto)session.getAttribute(USER), service.findByName(course));
+            } catch (EntityNotFoundException e) {
+                redirectAttributes.addFlashAttribute(MESSAGE, "User not on the course.");
+            }
         }
         return REDIRECT_COURSES;
     }
 
     @PostMapping("/update")
-    public String update(@ModelAttribute(COURSE) CourseDto course,
+    public String update(@Valid @ModelAttribute(COURSE) CourseDto course,
+                         BindingResult bindingResult,
                          RedirectAttributes redirectAttributes) {
-        try {
-            service.update(course);
-        } catch (EntityNotFoundException e) {
-            redirectAttributes.addFlashAttribute(MESSAGE, "Course not found");
+        if (bindingResult.hasErrors()) {
+            redirectAttributes.addFlashAttribute(MESSAGE,
+                    bindingResult.getAllErrors().get(0).getDefaultMessage());
+        } else {
+            try {
+                service.update(course);
+            } catch (EntityNotFoundException e) {
+                redirectAttributes.addFlashAttribute(MESSAGE, "Course not found");
+            }
         }
         return REDIRECT_COURSES;
     }
